// src/domain/models/facebook-account.ts
type FacebookData = {
  name: string;
  email: string;
  facebookId: string;
};
type AccountData = {
  name?: string;
  id?: string;
};
export class FacebookAccount {
  id?: string;
  name: string;
  email: string;
  facebookId: string;

  constructor(fbData: FacebookData, accountData?: AccountData) {
    this.id = accountData?.id;
    this.name = accountData?.name ?? fbData.name;
    this.email = fbData.email;
    this.facebookId = fbData.facebookId;
  }
}